Add SpriteSheet tests that cover drawFrame bug.
Adding tests that show a couple of different ways to initialize a SpriteSheet, and some tests showing that drawFrame successfully maps to a p5.image call given a few different ways to call it.  One of those ways is by frame index, which was the bug fixed by #90.  I undid that fix and confirmed that the new tests would have caught the bug.