Upgrade ESLint dependency
Fixes #93.

Update from ">=2.5.3 <2.10.0" to "^3.1.0" which is the latest version.  The major breaking change with this update is [dropping support for Node.js < 4](https://github.com/eslint/eslint/blob/master/docs/user-guide/migrating-to-3.0.0.md#dropping-support-for-nodejs--4) which shouldn't be an issue for this repo (as ESLint points out, Node 4 is the current LTS version).  I've added the engines field in package.json to document the dependency on Node >= 4.

Fixes the one new linting violation caught after the upgrade (and I'm actually surprised this one wasn't caught _before_ it).