Test to cover subtle bug in bad argument behavior
Realized that, as written, this method may affect play state even though you specified a frame that was out of bounds.  It looks like the intended behavior is for the call to be a no-op if the frame is out of bounds - after all, we won't actually be playing toward the specified frame.