Make examples page load p5.play directly
It seems registering an init method was not working on the examples page because it would load p5.js once on the page, but would eval p5.play with each example - and also because, for some reason, it was clearing all registered methods prior to eval'ing the examples.  I verified that removing this does not cause us to register init too many times.

One nice side effect of this is that you can debug p5.play on the examples page now, which was previously difficult since it was being eval'd.

I think the biggest risk is possible state bleed between examples, but if that's possible I haven't seen it happen yet.  We seem to create a new p5 instance for every example anyway.