Fix spritesheets
Fixes #125, supercedes #126.

This bug arises from a bizarre behavior in the `Animation` class, which can be constructed in (at least) three different ways:

* A: With a list of images `new Animation(pInst, image1, image2, image3)`
* B: With a range of images `new Animation(pInst, 'image0001.png', 'image0005.png')`
* C: With a spritesheet `new Animation(pInst, spritesheet)`

Unfortunately the internal structure of the `Animation` changes depending on how it was constructed.  In cases A and B, `this.images` is an array of `p5.Image` objects.  In case C, `this.images` is an array of POJOs in the form `{name: <string|number>, frame: {x: <number>, y: <number>, width: <number>, height: <number>}}`.

This becomes a problem when, for example, you want the width of the current frame.  In cases A and B you would ask for `this.images[current].width`, but in case C you have to ask for `this.images[current].frame.width`.

The change in #99 that introduced the spritesheet regression was a first attempt to unify these internal models, but they obviously missed a lot.  This change puts things back, fixes width/height accesses, and adds some basic tests for the interanl structure of `Animation` in different cases just to demonstrate the odd behavior.